update otel collector clusterrole
